requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1  # ?
Pillow==9.1.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
psycopg2-binary==2.9.3
sqlparse==0.4.4
pytz==2023.3
python-dotenv==1.0.0
django-debug-toolbar==4.2.0
